backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
python-jose[cryptography]==3.3.0
